feat: 도서상태 등록 repository 테스트 추가

+++ src/main/java/shop/nuribooks/books/dto/book/BookRegisterReq.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class BookRegisterReq {
 	@NotNull
-	private Long stateId;
+	private Integer stateId;
 
 	@NotNull
 	private Long publisherId;
